:recycle: Extract type_id.go
